requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
pydantic==2.6.1
pydantic-settings==2.1.0 